Feat-5 move tests to correct location
